+++ src/debug.hpp
@@ -9,7 +9,7 @@ LULU_FUNC int
 debug_get_pad(const Chunk *c);
 
 LULU_FUNC void
-debug_disassemble_at(const Chunk *c, int pc, int pad);
+debug_disassemble_at(const Chunk *c, isize pc, int pad);
 
 
 /**
@@ -20,4 +20,4 @@ debug_disassemble_at(const Chunk *c, int pc, int pad);
  *      The index of the instruction where `local_number` is valid.
  */
 LULU_FUNC const char *
-debug_get_local(const Chunk *c, int local_number, int pc);
+debug_get_local(const Chunk *c, int local_number, isize pc);
